requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bootstrap4==0.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
Django==5.1.2
django-bootstrap4==24.4
django-stdimage==6.0.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.16.1
fonttools==4.55.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.2.0
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.1
tiktoken==0.8.0
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.0
